refactor(ui): Separate entity lineage counts query from rest of entity query (#7569)
