+++ datahub-web-react/src/app/entity/shared/containers/profile/nav/ProfileNavBrowsePath.tsx
@@ -1,57 +1,12 @@
 import React from 'react';
-import { Link, useHistory, useLocation } from 'react-router-dom';
-import { Badge, Breadcrumb, Row } from 'antd';
+import { Link } from 'react-router-dom';
 import styled from 'styled-components/macro';
-import { InfoCircleOutlined, PartitionOutlined } from '@ant-design/icons';
-import { grey, blue } from '@ant-design/colors';
+import { Breadcrumb, Row } from 'antd';
 import { EntityType } from '../../../../../../types.generated';
 import { useEntityRegistry } from '../../../../../useEntityRegistry';
 import { PageRoutes } from '../../../../../../conf/Global';
-import { navigateToLineageUrl } from '../../../../../lineage/utils/navigateToLineageUrl';
-import useIsLineageMode from '../../../../../lineage/utils/useIsLineageMode';
-import { ANTD_GRAY, ENTITY_TYPES_WITH_MANUAL_LINEAGE } from '../../../constants';
-import { useGetLineageTimeParams } from '../../../../../lineage/utils/useGetLineageTimeParams';
-
-type Props = {
-    type: EntityType;
-    path: Array<string>;
-    upstreams: number;
-    downstreams: number;
-    breadcrumbLinksEnabled: boolean;
-};
-
-const LineageIconGroup = styled.div`
-    width: 180px;
-    display: flex;
-    justify-content: space-between;
-    margin-right: 8px;
-`;
-
-const LineageIcon = styled(PartitionOutlined)`
-    font-size: 20px;
-    vertical-align: middle;
-    padding-right: 6px;
-`;
-
-const DetailIcon = styled(InfoCircleOutlined)`
-    font-size: 20px;
-    vertical-align: middle;
-    padding-right: 6px;
-`;
-
-const IconGroup = styled.div<{ isSelected: boolean; disabled?: boolean }>`
-    font-size: 14px;
-    color: ${(props) => {
-        if (props.disabled) {
-            return grey[2];
-        }
-        return !props.isSelected ? 'black' : props.theme.styles['primary-color'] || blue[4];
-    }};
-    &:hover {
-        color: ${(props) => (props.disabled ? grey[2] : props.theme.styles['primary-color'] || blue[4])};
-        cursor: ${(props) => (props.disabled ? 'not-allowed' : 'pointer')};
-    }
-`;
+import { ANTD_GRAY } from '../../../constants';
+import { LineageSelector } from './LineageSelector';
 
 const BrowseRow = styled(Row)`
     padding: 10px 20px;
@@ -61,49 +16,24 @@ const BrowseRow = styled(Row)`
     border-bottom: 1px solid ${ANTD_GRAY[4.5]};
 `;
 
-const LineageNavContainer = styled.div`
-    display: inline-flex;
-    line-height: 24px;
-    align-items: center;
-`;
-
-const LineageSummary = styled.div`
-    margin-left: 16px;
-`;
-
-const LineageBadge = styled(Badge)`
-    &&& .ant-badge-count {
-        background-color: ${ANTD_GRAY[1]};
-        color: ${ANTD_GRAY[9]};
-        border: 1px solid ${ANTD_GRAY[5]};
-        font-size: 12px;
-        font-weight: 600;
-        height: 22px;
-    }
-`;
-
 export const BreadcrumbItem = styled(Breadcrumb.Item)<{ disabled?: boolean }>`
     &&& :hover {
         color: ${(props) => (props.disabled ? ANTD_GRAY[7] : props.theme.styles['primary-color'])};
     }
 `;
 
+type Props = {
+    urn: string;
+    type: EntityType;
+    path: Array<string>;
+    breadcrumbLinksEnabled: boolean;
+};
+
 /**
  * Responsible for rendering a clickable browse path view.
  */
-// TODO(Gabe): use this everywhere
-export const ProfileNavBrowsePath = ({
-    type,
-    path,
-    upstreams,
-    downstreams,
-    breadcrumbLinksEnabled,
-}: Props): JSX.Element => {
+export const ProfileNavBrowsePath = ({ urn, type, path, breadcrumbLinksEnabled }: Props): JSX.Element => {
     const entityRegistry = useEntityRegistry();
-    const history = useHistory();
-    const location = useLocation();
-    const isLineageMode = useIsLineageMode();
-    const { startTimeMillis, endTimeMillis } = useGetLineageTimeParams();
 
     const createPartialPath = (parts: Array<string>) => {
         return parts.join('/');
@@ -129,12 +59,6 @@ export const ProfileNavBrowsePath = ({
         </BreadcrumbItem>
     ));
 
-    const hasLineage = upstreams > 0 || downstreams > 0;
-    const canNavigateToLineage = hasLineage || ENTITY_TYPES_WITH_MANUAL_LINEAGE.has(type);
-
-    const upstreamText = upstreams === 100 ? '100+' : upstreams;
-    const downstreamText = downstreams === 100 ? '100+' : downstreams;
-
     return (
         <BrowseRow>
             <Breadcrumb style={{ fontSize: '16px' }} separator=">">
@@ -149,49 +73,7 @@ export const ProfileNavBrowsePath = ({
                 </BreadcrumbItem>
                 {pathCrumbs}
             </Breadcrumb>
-            <LineageNavContainer>
-                <LineageIconGroup>
-                    <IconGroup
-                        disabled={!canNavigateToLineage}
-                        isSelected={!isLineageMode}
-                        onClick={() => {
-                            if (canNavigateToLineage) {
-                                navigateToLineageUrl({
-                                    location,
-                                    history,
-                                    isLineageMode: false,
-                                    startTimeMillis,
-                                    endTimeMillis,
-                                });
-                            }
-                        }}
-                    >
-                        <DetailIcon />
-                        Details
-                    </IconGroup>
-                    <IconGroup
-                        disabled={!canNavigateToLineage}
-                        isSelected={isLineageMode}
-                        onClick={() => {
-                            if (canNavigateToLineage) {
-                                navigateToLineageUrl({
-                                    location,
-                                    history,
-                                    isLineageMode: true,
-                                    startTimeMillis,
-                                    endTimeMillis,
-                                });
-                            }
-                        }}
-                    >
-                        <LineageIcon />
-                        Lineage
-                    </IconGroup>
-                </LineageIconGroup>
-                <LineageSummary>
-                    <LineageBadge count={`${upstreamText} upstream, ${downstreamText} downstream`} />
-                </LineageSummary>
-            </LineageNavContainer>
+            <LineageSelector urn={urn} type={type} />
         </BrowseRow>
     );
 };
