+++ datahub-web-react/src/graphql/chart.graphql
@@ -64,12 +64,6 @@ query getChart($urn: String!) {
         parentContainers {
             ...parentContainersFields
         }
-        upstream: lineage(input: { direction: UPSTREAM, start: 0, count: 100 }) {
-            ...partialLineageResults
-        }
-        downstream: lineage(input: { direction: DOWNSTREAM, start: 0, count: 100 }) {
-            ...partialLineageResults
-        }
         status {
             removed
         }
