+++ datahub-web-react/src/graphql/dataset.graphql
@@ -115,12 +115,6 @@ fragment nonSiblingDatasetFields on Dataset {
         timestampMillis
         lastUpdatedTimestamp
     }
-    upstream: lineage(input: { direction: UPSTREAM, start: 0, count: 100 }) {
-        ...partialLineageResults
-    }
-    downstream: lineage(input: { direction: DOWNSTREAM, start: 0, count: 100 }) {
-        ...partialLineageResults
-    }
     ...viewProperties
     autoRenderAspects: aspects(input: { autoRenderOnly: true }) {
         aspectName
