+++ datahub-web-react/src/graphql/mlFeature.graphql
@@ -1,12 +1,6 @@
 query getMLFeature($urn: String!) {
     mlFeature(urn: $urn) {
         ...nonRecursiveMLFeature
-        upstream: lineage(input: { direction: UPSTREAM, start: 0, count: 100 }) {
-            ...partialLineageResults
-        }
-        downstream: lineage(input: { direction: DOWNSTREAM, start: 0, count: 100 }) {
-            ...partialLineageResults
-        }
         featureTables: relationships(input: { types: ["Contains"], direction: INCOMING, start: 0, count: 100 }) {
             ...fullRelationshipResults
         }
