+++ datahub-web-react/src/graphql/mlModelGroup.graphql
@@ -21,11 +21,5 @@ query getMLModelGroup($urn: String!) {
         ) {
             ...fullRelationshipResults
         }
-        upstream: lineage(input: { direction: UPSTREAM, start: 0, count: 100 }) {
-            ...partialLineageResults
-        }
-        downstream: lineage(input: { direction: DOWNSTREAM, start: 0, count: 100 }) {
-            ...partialLineageResults
-        }
     }
 }
